Implement capitalize command
