--- packages/plugin-core/src/commands/Capitalize.ts
@@ -1,19 +1,0 @@
-import { DENDRON_COMMANDS } from "../constants";
-import { BasicCommand } from "./base";
-
-type CommandOpts = {};
-
-type CommandInput = {};
-
-type CommandOutput = void;
-
-export class CapitalizeCommand extends BasicCommand<
-  CommandOpts,
-  CommandOutput
-> {
-  key = DENDRON_COMMANDS.CAPITALIZE.key;
-
-  async execute() {
-    //TODO: implement command logic
-  }
-}
